Move expand config map converter back out of internal package
