+++ service/config_provider.go
@@ -27,7 +27,6 @@ import (
 	"go.opentelemetry.io/collector/config/configmapprovider"
 	"go.opentelemetry.io/collector/config/configunmarshaler"
 	"go.opentelemetry.io/collector/config/experimental/configsource"
-	"go.opentelemetry.io/collector/service/internal/configprovider"
 )
 
 // ConfigProvider provides the service configuration.
@@ -112,7 +111,7 @@ func NewDefaultConfigProvider(configFileName string, properties []string) Config
 		},
 		[]ConfigMapConverterFunc{
 			configmapprovider.NewOverwritePropertiesConverter(properties),
-			configprovider.NewExpandConverter(),
+			configmapprovider.NewExpandConverter(),
 		},
 		configunmarshaler.NewDefault())
 }
